feat: move to typescript
